feat: update pipeline event function

+++ lib/model/user/user.go
@@ -6,8 +6,8 @@ import "github.com/megakuul/battleshiper/lib/model/rbac"
 const USER_COLLECTION = "users"
 
 type ExecutionLimitCounter struct {
-	ExpirationTime     int `bson:"expiration_time"`
-	PipelineExecutions int `bson:"pipeline_executions"`
+	ExpirationTime     int64 `bson:"expiration_time"`
+	PipelineExecutions int   `bson:"pipeline_executions"`
 }
 
 type Repository struct {
@@ -17,7 +17,7 @@ type Repository struct {
 }
 
 type GithubData struct {
-	InstallationId int          `bson:"installation_id"`
+	InstallationId int64        `bson:"installation_id"`
 	Repositories   []Repository `bson:"repositories"`
 }
